cpufreq: loose TouchWiz freq min/max limit
TouchWiz often limits cluster1 cpufreq and that function calls hmp_boost,
It's not good for battery life.

But we also want TouchWiz to tell kernel what is user interface doing now.
So we can make cpufreq limitations more flexible, That make it better for
battery life but also reduce the impaction of user experience.

Signed-off-by: DevriesL <therkduan@gmail.com>